+++ src/als-tools.core/ValueObjects/TrackDelay.cs
@@ -6,7 +6,7 @@ public class TrackDelay
     /// The delay value/amount, either in ms or samples, depending on the
     /// <see cref="IsValueSampleBased"/> property.
     /// </summary>
-    public int? Value { get; set; }
+    public double? Value { get; set; }
 
     /// <summary>
     /// Whether the delay is sample based of not (in this case, milliseconds)
